+++ routes/index.js
@@ -1,36 +1,28 @@
+/**
+ * This router handles all the page loading. (i.e going from home page to bracket page)
+ */
+
 const express = require("express");
-const { route } = require("express/lib/application");
-const router = express.Router();
-const mongo = require('mongodb');
-const bodyParser = require("body-parser");
+const urlRouter = express.Router();
 
-router.use(bodyParser.urlencoded({ extended: false }));
-router.use(express.json());
-
-const db = 'mongodb+srv://me:1206825Rs!!!@futbolcluster.k7r86.mongodb.net/FutbolCluster?retryWrites=true&w=majority';
-
-var url = 'mongodb+srv://me:1206825Rs!!!@futbolcluster.k7r86.mongodb.net/FutbolCluster?retryWrites=true&w=majority';
-
-// this will be used in index.js
-router.get('/', (req, res, next) => {
-    // res.render(('index'));
-    res.sendFile(path.join(__dirname, 'views', 'index.html'));
+// This is for the bracket button, when this is clicked it will redirect to the bracket page
+urlRouter.get("/bracket", (req, res) => {
+    res.render('bracket');
 });
 
-router.post('/', function(req, res, next) {
-    var item = {
-        name: req.body.name,
-        email: req.body.email,
-        password: req.body.password,
-        password_confirmation: req.body.password_confirmation
-    };
-
-    db.collection("usermodels").insertOne(item, function(err, result) {
-        console.log("item inserted");
-        db.close();
-    });
-
-    res.redirect('/')
+// If you're outside the home page, this will redirect you back to the home page
+urlRouter.get("/index", (req, res) => {
+    res.render('index')
 });
 
-module.exports = router;
+// This is the route that will take us to the 2014 worldcup page
+urlRouter.get("/WorldCup2014", (req, res) => {
+    res.render('WorldCup2014')
+});
+
+// This is the route that will take us to the 2018 worldcup page
+urlRouter.get("/WorldCup2018", (req, res) => {
+    res.render('WorldCup2018')
+});
+
+module.exports = urlRouter;
