+++ server.js
@@ -1,8 +1,10 @@
 const mongoose = require('mongoose');
 const express = require("express");
-const path = require('path');
 const { check, validationResult } = require('express-validator');
-const users = require('./models/users');
+const urlRouter = require('./routes/index');
+const signUpRouter = require('./routes/signUp');
+// const path = require('path');
+// const users = require('./models/users');
 
 const app = express();
 const port = 5501;
@@ -10,8 +12,9 @@ const bodyParser = require("body-parser");
 
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(express.json());
+app.use('/', urlRouter);
+app.use('/signUp', signUpRouter);
 app.set('view engine', 'ejs');
-// app.use('/index', index);
 
 
 // This is what gets us connected to the mongoDB: futbolcluster
@@ -32,25 +35,6 @@ app.get("/", (req, res) => {
     res.render('index');
 });
 
-// This is for the bracket button, when this is clicked it will redirect to the bracket page
-app.get("/bracket", (req, res) => {
-    res.render('bracket');
-});
-
-// If you're outside the home page, this will redirect you back to the home page
-app.get("/index", (req, res) => {
-    res.render('index')
-});
-
-// This is the route that will take us to the 2014 worldcup page
-app.get("/WorldCup2014", (req, res) => {
-    res.render('WorldCup2014')
-});
-
-// This is the route that will take us to the 2018 worldcup page
-app.get("/WorldCup2018", (req, res) => {
-    res.render('WorldCup2018')
-});
 
 let options = {
     dotfiles: "ignore", // allow, deny, ignore
@@ -81,48 +65,6 @@ app.post('/signIn', [check('email').isEmail().normalizeEmail(),
         res.redirect('/');
     }
 );
-
-// This post will add a user 
-app.post('/signUp', [check('name', 'Username cannot be empty!').exists().isLength({ min: 3 }),
-        check('email')
-        .isEmail()
-        .normalizeEmail(),
-        check('password', 'Password must contain at least 8 characters')
-        .exists()
-        .isLength({ min: 8 })
-        .custom((val, { req, loc, path }) => {
-            if (val != req.body.password_confirmation) {
-                throw new Error("Password does not match.");
-            } else {
-                return validationResult;
-            }
-        }),
-        check('password_confirmation', 'Password does not match')
-        .exists()
-        .isLength({ min: 8 })
-    ],
-    function(req, res, next) {
-        var errors = validationResult(req)
-
-        if (!errors.isEmpty()) {
-            const alert = errors.array();
-            res.render('index', { alert });
-        } else {
-            var user = {
-                name: req.body.name,
-                email: req.body.email,
-                password: req.body.password,
-                password_confirmation: req.body.password_confirmation
-            };
-
-            db.collection("users").insertOne(user, function(err, result) {
-                console.log("item inserted");
-                db.close();
-            });
-
-            res.redirect('/')
-        }
-    });
 
 app.listen(port, function() {
     console.log('Server is running on port ' + port);
